outsourced AttackActions
